:sparkles: feat: rota /ativos e /conta funcionando

+++ src/database/migrations/20220724144456-create-conta.js
@@ -10,7 +10,7 @@ module.exports = {
         primaryKey: true,
       },
       Saldo: {
-        type: Sequelize.DECIMAL
+        type: Sequelize.DOUBLE
       }
     });
   },
